Agregado el Main, onResume en Login, seguir probando caminos de error
