+++ mobile/app/src/main/AndroidManifest.xml
@@ -27,13 +27,16 @@
         android:hardwareAccelerated="true"
         android:largeHeap="true">
 
-        <activity android:name=".login.LoginActivity">
-        </activity>
-        <activity android:name=".home.MainActivity" android:theme="@style/SplashTheme">
+        <activity android:name=".session.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".login.LoginActivity">
+        </activity>
+        <activity android:name=".home.MyLookActivity" android:theme="@style/SplashTheme">
+
         </activity>
         <activity android:name=".info.ArticleInfoActivity"></activity>
         <activity android:name=".storeProfile.StoreActivity"></activity>
