+++ mobile/app/src/main/java/com/mylook/mylook/home/MyLookActivity.java
@@ -1,27 +1,22 @@
 package com.mylook.mylook.home;
 
 import android.annotation.SuppressLint;
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.MenuItem;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
 import com.mylook.mylook.R;
 import com.mylook.mylook.closet.ClosetFragment;
 import com.mylook.mylook.explore.ExploreFragment;
-import com.mylook.mylook.login.LoginActivity;
 import com.mylook.mylook.profile.ProfileFragment;
 import com.mylook.mylook.recommend.RecommendFragment;
 import com.mylook.mylook.session.Sesion;
 
-public class MainActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener {
+public class MyLookActivity extends AppCompatActivity implements BottomNavigationView.OnNavigationItemSelectedListener {
     private Toolbar toolbar;
     private String activeFragment;
     Fragment fragment = null;
@@ -32,34 +27,14 @@ public class MainActivity extends AppCompatActivity implements BottomNavigationV
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         activeFragment = HomeFragment.TAG;
-        getSession();
-        if(currentSesion!=null) {
-            currentSesion.initializeElements().addOnCompleteListener(new OnCompleteListener() {
-                @Override
-                public void onComplete(@NonNull Task task) {
-                    loadFragment(HomeFragment.getInstance());
-                    setContentView(R.layout.activity_main);
-                    BottomNavigationView navigation = findViewById(R.id.navigation);
-                    navigation.setOnNavigationItemSelectedListener(MainActivity.this);
-                    toolbar = findViewById(R.id.main_toolbar);
-                    setSupportActionBar(toolbar);
-                    setTheme(R.style.AppTheme);
-                }
-            });
-        }else{
-            Intent intent = new Intent(MainActivity.this, LoginActivity.class);
-            startActivity(intent);
-            try {
-                finalize();
-            }catch (Throwable e){
-                Log.e("current Sesion is null ", e.getMessage() );
 
-            }
-        }
-    }
-
-    private void getSession(){
-        currentSesion = Sesion.getInstance();
+        loadFragment(HomeFragment.getInstance());
+        setContentView(R.layout.activity_mylook_app);
+        BottomNavigationView navigation = findViewById(R.id.navigation);
+        navigation.setOnNavigationItemSelectedListener(MyLookActivity.this);
+        toolbar = findViewById(R.id.main_toolbar);
+        setSupportActionBar(toolbar);
+        setTheme(R.style.AppTheme);
     }
