+++ mobile/app/src/main/res/layout/activity_mylook_app.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.MainActivity">
+    tools:context=".home.MyLookActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/main_toolbar"
