+++ mobile/app/src/main/res/layout/activity_store_final.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_gravity="fill_vertical"
-    tools:context=".home.MainActivity">
+    tools:context=".home.MyLookActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
